Move the standard close action binding to MenuCaptionBarForm.
